+++ src/main/java/oit/is/hondaken/scheduler/controller/ClassController.java
@@ -177,13 +177,5 @@ public class ClassController {
   }
 
 
-  @GetMapping("step2")
-  public String step2() {
-    return "zikken2.html";
-  }
 
-  @PostMapping("step2")
-  public String reStep2() {
-    return "index.html";
-  }
 }
