+++ src/main/resources/templates/timeout.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+  <title>セッションタイムアウト</title>
+</head>
+
+<body>
+  <h1>セッションがタイムアウトしました</h1>
+  <p>再度ログインしてください。</p>
+  <a href="/index.html">TOPへ戻る</a>
+</body>
+
+</html>
